backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78
albumentations==1.3.1
numpy==1.26.2

# Background tasks
celery==5.3.4
redis==5.0.1

# Object storage
boto3==1.29.7

# Export formats
pycocotools==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
